requirements.txt
mysql-connector-python>=8.0.33, <8.1.0
grpcio==1.54.0
protobuf==3.20.3
pycryptodome>=3.17